fix: CoinItem component 분리
