+++ src/components/CoinItem/Price/index.tsx
